+++ src/components/Typography/CHANGELOG.md
@@ -7,3 +7,6 @@
 - 2025-10-05: Add `Typography.Ul` and `Typography.Li` components. Also add
   default styles for `Typography.P`.
 - 2025-10-10: Add `Typography.Code` component for inline code.
+- 2025-10-12: Add `Typography.Root` component as a wrapper for all typography
+  elements. Improve overall styles and spacing. Put styles into `components` CSS
+  layer.